refactor: Add test_main_test.go for adminclienthandlers package
Add a new file test_main_test.go to the adminclienthandlers package. This file contains the TestMain function that initializes the configuration and runs the tests for the package.

refactor: Update HandleAdminClientGenerateNewSecretGet function

In the handler_admin_client_authentication.go file, update the HandleAdminClientGenerateNewSecretGet function. Instead of using a struct to store the result, use a map to store the new secret generated using the stringutil.GenerateSecurityRandomString function.