Refactor documentation and development environment setup
Simplify and clarify the Goiabada documentation by rephrasing and organizing the user management features. Additionally, update the development environment setup by installing mkdocs and mkdocs-material for better documentation generation.

- Update user management description
- Improve authentication and authorization explanation
- Enhance control over resources and permissions
- Add self-service account page for users
- Add OAuth2 and OpenID Connect support
- Prepare for compatibility with MySQL and SQLite databases

Refactor Dockerfile to use Go 1.23.2-alpine as buildStage

Update the Dockerfile in the .devcontainer directory to use the golang:1.23.2-alpine image as the build stage. This change improves the speed and reduces the resource consumption of the Goiabada container.

Refactor account_phone.html to update phone update message

Update the account_phone.html template in the adminconsole/web/template directory to improve the clarity of the phone update message. The change removes the mention of phone verification and focuses solely on updating the phone number.